refactor(constants and types): created config/types.ts and core/constants.ts to better group interfaces and constants + provide easier UI access to sifting rules later
