+++ src/config/types.ts
@@ -1,0 +1,10 @@
+export interface MatchGroup {
+    exact: string[];
+    partial: string[];
+}
+
+export interface SifterConfig {
+    title: MatchGroup;
+    company: MatchGroup;
+    location: MatchGroup;
+}
