+++ src/content/main.ts
@@ -1,5 +1,5 @@
 import { JobSifter } from "../core/JobSifter";
-import { SIFTER_KEYWORDS, SIFTER_CONFIG } from "../config/constants";
+import { SIFTER_CONFIG } from "../config/constants";
 
 const goldpan = new JobSifter(SIFTER_CONFIG);
