+++ src/core/constants.ts
@@ -1,0 +1,4 @@
+export const SELECTORS = {
+    JOB_CONTAINER: '.scaffold-layout__list',
+    JOB_CARD: 'li[data-occludable-job-id]',
+} as const
