+++ src/core/types.ts
@@ -5,14 +5,3 @@ export interface JobCardData {
     location: string;
     //add more fields as needed
 }
-
-export interface MatchGroup {
-    exact: string[];
-    partial: string[];
-}
-
-export interface SifterConfig {
-    title: MatchGroup;
-    company: MatchGroup;
-    location: MatchGroup;
-}
